Migrate bin/cli.js to TypeScript
- Move CLI logic from bin/cli.js to src/cli.ts with type annotations
- bin/cli.js becomes a thin shebang wrapper: require('../dist/cli')
- CLI dependencies (chalk, commander, liftoff, minimist, tildify) are
  already at latest CJS-compatible versions (chalk 5.x and tildify 3.x
  are ESM-only, so 4.x and 2.x are retained respectively)
